requirements.txt
glustercli==0.3
